fix : 채팅창 컴포넌트 작동 안되는 오류 수정
